requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
# Editable Git install with no remote (preshow_mailer==0.1.0)
-e /Users/Work/preshow_mailer
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
urllib3==2.5.0
